@@ -2,13 +2,11 @@ using System.Security.Cryptography;
 
 namespace Tactical.DDD.Tests.TestAggregate
 {
-    public sealed class SubTask : Entity<TaskId>
+    public sealed class SubTask : Entity<SubTaskId>
     {
-        public override TaskId Id { get; protected set; }
-
         public string Title { get; private set; }
         
-        public SubTask(TaskId id, string title)
+        public SubTask(SubTaskId id, string title)
         {
             Id = id;
             Title = title;
