@@ -1,15 +1,7 @@
-using System;
-using System.Collections.Generic;
-
 namespace Tactical.DDD
 {
-    public abstract class EntityId : ValueObject, IEntityId
+    public abstract record EntityId
     {
-        public abstract override string ToString();
-
-        protected override IEnumerable<object> GetAtomicValues()
-        {
-            yield return ToString();
-        }
+        public static implicit operator string(EntityId id) => id.ToString();
     }
 }
